add intermediete abstract to mutators

+++ include/drawing/openGL/shaders/Triangle2.cpp
@@ -1,5 +1,6 @@
 #include "Triangle2.hpp"
 #include "common/Config.hpp"
+#include "my_utils/Profiler.hpp"
 #include <cstddef>
 
 namespace shaders {
@@ -67,11 +68,11 @@ int Triangle2::bindDataToBuffer(Genotype& genes)
     int i = 0;
     for (auto gene : genes.genes) {
         glm::vec4 color = { gene.color.r, gene.color.g, gene.color.b, gene.color.a };
-        float distance = 1.f / max;
+        float distance = 1.f / (float)max;
 
         vInputs.emplace_back(
             0,
-            distance * i - 1.f,
+            distance * (float)i - 1.f,
             glm::vec2(gene.position.x, gene.position.y),
             0,
             color //
@@ -79,7 +80,7 @@ int Triangle2::bindDataToBuffer(Genotype& genes)
 
         vInputs.emplace_back(
             gene.scale.x,
-            distance * i - 1.f,
+            distance * (float)i - 1.f,
             glm::vec2(gene.position.x, gene.position.y),
             gene.rotation * 3.14 * 2,
             color //
@@ -87,7 +88,7 @@ int Triangle2::bindDataToBuffer(Genotype& genes)
 
         vInputs.emplace_back(
             gene.scale.y,
-            distance * i - 1.f,
+            distance * (float)i - 1.f,
             glm::vec2(gene.position.x, gene.position.y),
             gene.rotation * 3.14 * 2 + 3.14 / 2,
             color //
